trying to upload

+++ app/src/main/java/com/jianzheng/studyzero/data/Response.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class Response(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
-    val delay: Int,
+    val delay: Long,
     val answer1: Int,
     val answer2: Int,
     val startingTime: Long,
